audit: memory safety audit — 42 findings resolved, 16 code fixes
Memory Safety Audit (March 2026) — all 42 findings verified and resolved:
- 16 FIXED with code changes across 9 pallets
- 20 FALSE POSITIVE (saturating/checked arithmetic already in use)
- 5 ACCEPTABLE (standard .expect() in node startup/tests)
- 1 NON-EXISTENT (finding does not match actual code)

Code fixes:
- belizex: F-1/F-2/F-3/F-4 — DEX math migrated to U256 intermediates
- belizex: BEX-NextOrderId — u32 wrapping handled with checked_add
- oracle: O-C-03 — overflow-safe median calculation
- oracle: O-INPUT-01 — reject zero-value price submissions
- oracle: O-C-02 — redundant deviation validation
- justice: JUST-01 — slash_bps bounded to 10_000
- staking: S-CRIT-2 — rolling average counter widened to u64
- governance: CAST-GOV-01 — .min(100) before u8 cast
- community: COMM-01/03/04 — saturating arithmetic for scores
- quantum: QTM-01 — royalty_rate validation (max 10_000 bps)
- payroll: PAY-04 — saturating payroll calculations
- compliance: ARITH-COMP-01 — saturating score update

Also includes:
- Dependency audit report and deny.toml (cargo-deny config)
- Security audit CI workflow
- Updated Cargo.toml and Cargo.lock

+++ pallets/staking/src/lib.rs
@@ -1183,11 +1183,12 @@ pub mod pallet {
                 // Update contribution count
                 stats.contribution_count = stats.contribution_count.saturating_add(1);
 
-                // Update rolling average quality
-                let total_quality = (stats.avg_quality as u32)
-                    .saturating_mul(stats.contribution_count.saturating_sub(1))
-                    .saturating_add(quality_score as u32);
-                stats.avg_quality = (total_quality / stats.contribution_count) as u8;
+                // Update rolling average quality (use u64 intermediate to prevent
+                // overflow when avg_quality * (count-1) exceeds u32::MAX at ~16.8M contributions)
+                let total_quality = (stats.avg_quality as u64)
+                    .saturating_mul(stats.contribution_count.saturating_sub(1) as u64)
+                    .saturating_add(quality_score as u64);
+                stats.avg_quality = (total_quality / stats.contribution_count as u64) as u8;
 
                 // Update total volume
                 stats.total_volume = stats.total_volume.saturating_add(volume_kb as u64);
